extract useIframeMessenger hook
